refactor: 패키지 정리, 로또 생성로직 streamify
