+++ src/main/java/step3/util/LottoNumberGenerator.java
@@ -1,4 +1,4 @@
-package step3;
+package step3.util;
 
 import java.util.List;
