+++ src/main/java/step3/view/LottoInputView.java
@@ -1,4 +1,4 @@
-package step3;
+package step3.view;
 
 import java.util.Arrays;
 import java.util.List;
