+++ src/test/java/step3/LottoBucketTest.java
@@ -10,7 +10,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 public class LottoBucketTest {
     @Test
     @DisplayName("LottoNumbers 추가 테스트")
-    void addNewLottoNumbersTest() {
+    void addTest() {
         LottoBucket lottoBucket = new LottoBucket();
 
         int before = lottoBucket.size();
@@ -20,5 +20,19 @@ public class LottoBucketTest {
         int after = lottoBucket.size();
 
         assertThat(after).isEqualTo(before + 1);
+    }
+
+    @Test
+    @DisplayName("LottoBucket간 합치는 테스트")
+    void mergeTest() {
+        LottoBucket lottoBucket1 = new LottoBucket();
+        lottoBucket1.add(new LottoEntry(Arrays.asList(1, 2, 3, 4, 5, 6)));
+
+        LottoBucket lottoBucket2 = new LottoBucket();
+        lottoBucket2.add(new LottoEntry(Arrays.asList(1, 2, 3, 4, 5, 6)));
+
+        lottoBucket1.merge(lottoBucket2);
+
+        assertThat(lottoBucket1.size()).isEqualTo(2);
     }
 }
