+++ src/test/java/step3/MatchStatisticsTest.java
@@ -2,6 +2,7 @@ package step3;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import step3.enums.Match;
 
 import java.util.List;
 import java.util.stream.Collectors;
